update some json request to rapidjson, move validation into transaction model

+++ login_server/src/cpp/JSONInterface/JsonAppLogin.cpp
@@ -1,58 +1,45 @@
 #include "JsonAppLogin.h"
 
-#include "Poco/URI.h"
-
-#include "../lib/DataTypeConverter.h"
-
 #include "../controller/AppAccessToken.h"
 #include "../controller/User.h"
 
 #include "../SingletonManager/SessionManager.h"
 
+#include "rapidjson/document.h"
 
-Poco::JSON::Object* JsonAppLogin::handle(Poco::Dynamic::Var params)
+using namespace rapidjson;
+
+Document JsonAppLogin::handle(const Document& params)
 {
 	Poco::UInt64 access_token_code;
-	if (params.isVector()) {
-		try {
-			const Poco::URI::QueryParameters queryParams = params.extract<Poco::URI::QueryParameters>();
-			for (auto it = queryParams.begin(); it != queryParams.end(); it++) {
-				if (it->first == "access_token") {
-					auto numberParseResult = DataTypeConverter::strToInt(it->second, access_token_code);
-					if (DataTypeConverter::NUMBER_PARSE_OKAY != numberParseResult) {
-						return stateError("error parsing access token", DataTypeConverter::numberParseStateToString(numberParseResult));
-					}
-					break;
-				}
-			}
-			//auto var = params[0];
-		}
-		catch (Poco::Exception& ex) {
-			return stateError("error parsing query params, Poco Error", ex.displayText());
-		}
+	auto access_token_result = getUInt64Parameter(params, "access_token", access_token_code);
+	if (access_token_result.IsObject()) {
+		return access_token_result;
 	}
+	
 	auto sm = SessionManager::getInstance();
 	auto access_token = controller::AppAccessToken::load(access_token_code);
 	if (access_token.isNull()) {
-		return stateError("access token not found");
+		return rstateError("access token not found");
 	}
 	Poco::Timespan max_age;
 	max_age.assign(7, 0, 0, 0, 0);
 	if (access_token->getAge() > max_age) {
 		access_token->deleteFromDB();
-		return stateError("access token to old");
+		return rstateError("access token to old");
 	}
 	access_token->getModel()->update();
 	auto session = sm->getNewSession();
 	auto user = controller::User::create();
 	if (1 != user->load(access_token->getModel()->getUserId())) {
-		return stateError("access token invalid");
+		return rstateError("access token invalid");
 	}
 	session->setUser(user);
 
-	Poco::JSON::Object* result = new Poco::JSON::Object;
-	result->set("state", "success");
-	result->set("session_id", session->getHandle());
+	Document result;
+	result.SetObject();
+	result.AddMember("state", "success", result.GetAllocator());
+	result.AddMember("session_id", session->getHandle(), result.GetAllocator());
 
 	return result;
